core/requirements.txt
# django 
django 
django-decouple
django-cors-headers 
django-modeltranslation 
django-mptt 
django-crispy-forms
django-allauth 
django-import-export
django-debug-toolbar


# django admin 
django-admin-tools 
django-admin-sortable
django-admin-sortable2
django-admin-rangefilter
django-admin-autocomplete-filter
django-autocomplete-light
django-nested-admin
django-ckeditor
django-tinymce4-lite 
django-filebrowser-no-grappelli 
django-grappelli 
django-colorfield
django-rosetta 


# RESTful API
django-rest-framework
djoser 
djangorestframework_simplejwt

# graphql


# python 
transliterate 
pillow 
selenium 
bs4 
xlsxwriter
pandas 
celery 


